Fix checkstyle; add new sec logic for OIDC ATs

+++ ci/tests/puppeteer/run.sh
@@ -22,13 +22,57 @@ function printred() {
 casVersion=(`cat $PWD/gradle.properties | grep "version" | cut -d= -f2`)
 echo -n "Running Puppeteer tests for Apereo CAS Server: " && printcyan "${casVersion}"
 
-scenario="$1"
+while (( "$#" )); do
+  case "$1" in
+  --scenario)
+    scenario="$2"
+    shift 2
+    ;;
+  --install-puppeteer)
+      INSTALL_PUPPETEER="true"
+      shift 1
+      ;;
+  --debug)
+    DEBUG="-debug"
+    shift 1
+    ;;
+  --debug-port)
+    DEBUG_PORT="$2"
+    shift 2
+    ;;
+  --debug-suspend)
+    DEBUG_SUSPEND="y"
+    shift 1
+    ;;
+  --rebuild)
+    REBUILD="true"
+    shift 1
+    ;;
+  --buildflags)
+    BUILDFLAGS="$2"
+    shift 2
+    ;;
+  --headless)
+    export HEADLESS="true"
+    shift 1;
+    ;;
+  --rerun)
+    RERUN="true"
+    shift 1;
+    ;;
+  *)
+    printred "Unknown flag: $1"
+    exit 1
+    ;;
+  esac
+done
+
 if [[ -z "$scenario" ]] ; then
-  printred "Usage: ./ci/tests/puppeteer/run.sh ${PWD}/ci/tests/puppeteer/scenarios/[scenario name]"
+  printred "Missing scenario name for the test"
   exit 1
 fi
 if [[ ! -d "${scenario}" ]]; then
-  printred "${scenario} doesn't exist."
+  printred "Scenario ${scenario} doesn't exist."
   exit 1;
 fi
 
@@ -36,20 +80,20 @@ scenarioName=${scenario##*/}
 export SCENARIO="${scenarioName}"
 
 if [[ "${CI}" == "true" ]]; then
-  DEBUG=${2}
-else
-  DEBUG=${2:-debug}
+  printgreen "DEBUG flag is turned off while running CI"
+  DEBUG=""
 fi
 
-DEBUG_PORT=${3:-5000}
-DEBUG_SUSPEND=${4:-n}
+if [[ "${RERUN}" == "true" ]]; then
+  REBUILD="false"
+fi
 
 #echo "Installing jq"
 #sudo apt-get install jq
 
 random=$(openssl rand -hex 8)
 
-if [[ ! -d "$PWD"/ci/tests/puppeteer/node_modules/puppeteer ]] ; then
+if [[ ! -d "$PWD"/ci/tests/puppeteer/node_modules/puppeteer || "${INSTALL_PUPPETEER}" == "true" ]]; then
   printgreen "Installing Puppeteer"
   cd "$PWD"/ci/tests/puppeteer
   npm_install_cmd="npm install"
@@ -59,19 +103,24 @@ else
   printgreen "Using existing Puppeteer modules..."
 fi
 
-echo "Creating overlay work directory"
-rm -Rf "$PWD"/ci/tests/puppeteer/overlay
-mkdir "$PWD"/ci/tests/puppeteer/overlay
+if [[ "${RERUN}" != "true" ]]; then
+  echo "Creating overlay work directory"
+  rm -Rf "$PWD"/ci/tests/puppeteer/overlay
+  mkdir "$PWD"/ci/tests/puppeteer/overlay
+fi
 
-dname="${dname:-CN=cas.example.org,OU=Example,OU=Org,C=US}"
-subjectAltName="${subjectAltName:-dns:example.org,dns:localhost,ip:127.0.0.1}"
 keystore="$PWD"/ci/tests/puppeteer/overlay/thekeystore
-printgreen "\nGenerating keystore ${keystore} for CAS with\nDN=${dname}, SAN=${subjectAltName} ..."
-[ -f "${keystore}" ] && rm "${keystore}"
-keytool -genkey -noprompt -alias cas -keyalg RSA -keypass changeit -storepass changeit \
-  -keystore "${keystore}" -dname "${dname}" 
-[ -f "${keystore}" ] && echo "Created ${keystore}"
 export CAS_KEYSTORE="${keystore}"
+
+if [[ "${RERUN}" != "true" ]]; then
+  dname="${dname:-CN=cas.example.org,OU=Example,OU=Org,C=US}"
+  subjectAltName="${subjectAltName:-dns:example.org,dns:localhost,ip:127.0.0.1}"
+  printgreen "\nGenerating keystore ${keystore} for CAS with\nDN=${dname}, SAN=${subjectAltName} ..."
+  [ -f "${keystore}" ] && rm "${keystore}"
+  keytool -genkey -noprompt -alias cas -keyalg RSA -keypass changeit -storepass changeit \
+    -keystore "${keystore}" -dname "${dname}"
+  [ -f "${keystore}" ] && echo "Created ${keystore}"
+fi
 
 echo -e "******************************************************"
 printgreen "Scenario: ${scenario}"
@@ -91,6 +140,7 @@ if [[ $project == starter* ]]; then
   projectType=jar
 fi
 
+
 casWebApplicationFile="${PWD}/webapp/cas-server-webapp-${project}/build/libs/cas-server-webapp-${project}-${casVersion}.${projectType}"
 if [[ ! -f "$casWebApplicationFile" ]]; then
   printyellow "CAS web application at ${casWebApplicationFile} cannot be found. Rebuilding..."
@@ -99,7 +149,7 @@ fi
 
 dependencies=$(cat "${config}" | jq -j '.dependencies')
 
-if [[ "${REBUILD}" != "false" ]]; then
+if [[ "${REBUILD}" == "true" && "${RERUN}" != "true" ]]; then
   printgreen "\nBuilding CAS found in $PWD for dependencies [${dependencies}] with flags [${BUILDFLAGS}]"
   ./gradlew :webapp:cas-server-webapp-${project}:build -DskipNestedConfigMetadataGen=true -x check -x javadoc \
     --no-daemon --build-cache --configure-on-demand --parallel -PcasModules="${dependencies}" -q ${BUILDFLAGS}
@@ -107,56 +157,61 @@ if [[ "${REBUILD}" != "false" ]]; then
     printred "\nFailed to build CAS web application. Examine the build output."
     exit 1
   fi
-else
-  printgreen "\nNot rebuilding CAS web application file because REBUILD=false"
-fi
-cp ${casWebApplicationFile} "$PWD"/cas.${projectType}
-if [ $? -eq 1 ]; then
-  printred "Unable to build or locate the CAS web application file. Aborting test..."
-  exit 1
 fi
 
-# paths passed as arguments on command line get converted
-# by msys2 on windows, in some cases that is good but not with --somearg=file:/${PWD}/ci/...
-# so this converts path to windows format
-PORTABLE_PWD=${PWD}
-command -v cygpath > /dev/null && test ! -z "$MSYSTEM"
-if [[ $? -eq 0 ]]; then
-  PORTABLE_PWD=$(cygpath -w $PWD)
+if [[ "${RERUN}" != "true" ]]; then
+  cp ${casWebApplicationFile} "$PWD"/cas.${projectType}
+  if [ $? -eq 1 ]; then
+    printred "Unable to build or locate the CAS web application file. Aborting test..."
+    exit 1
+  fi
 fi
 
-initScript=$(cat "${config}" | jq -j '.initScript // empty')
-initScript="${initScript//\$\{PWD\}/${PWD}}"
-initScript="${initScript//\$\{SCENARIO\}/${scenarioName}}"
-[ -n "${initScript}" ] && \
-  printgreen "Initialization script: ${initScript}" && \
-  chmod +x "${initScript}" && \
-  eval "export SCENARIO=${scenarioName}"; eval "${initScript}"
 
-runArgs=$(cat "${config}" | jq -j '.jvmArgs // empty')
-runArgs="${runArgs//\$\{PWD\}/${PWD}}"
-[ -n "${runArgs}" ] && echo -e "JVM runtime arguments: [${runArgs}]"
+if [[ "${RERUN}" != "true" ]]; then
+  # Paths passed as arguments on command line get converted
+  # by msys2 on windows, in some cases that is good but not with --somearg=file:/${PWD}/ci/...
+  # so this converts path to windows format
+  PORTABLE_PWD=${PWD}
+  command -v cygpath > /dev/null && test ! -z "$MSYSTEM"
+  if [[ $? -eq 0 ]]; then
+    PORTABLE_PWD=$(cygpath -w $PWD)
+  fi
 
-properties=$(cat "${config}" | jq -j '.properties // empty | join(" ")')
-properties="${properties//\$\{PWD\}/${PORTABLE_PWD}}"
-properties="${properties//\$\{SCENARIO\}/${scenarioName}}"
-properties="${properties//\%\{random\}/${random}}"
-if [[ "$DEBUG" == "debug" ]]; then
-  printyellow "Enabling debugger on port $DEBUG_PORT"
-  runArgs="${runArgs} -Xrunjdwp:transport=dt_socket,address=$DEBUG_PORT,server=y,suspend=$DEBUG_SUSPEND"
+  initScript=$(cat "${config}" | jq -j '.initScript // empty')
+  initScript="${initScript//\$\{PWD\}/${PWD}}"
+  initScript="${initScript//\$\{SCENARIO\}/${scenarioName}}"
+  [ -n "${initScript}" ] && \
+    printgreen "Initialization script: ${initScript}" && \
+    chmod +x "${initScript}" && \
+    eval "export SCENARIO=${scenarioName}"; eval "${initScript}"
+
+  runArgs=$(cat "${config}" | jq -j '.jvmArgs // empty')
+  runArgs="${runArgs//\$\{PWD\}/${PWD}}"
+  [ -n "${runArgs}" ] && echo -e "JVM runtime arguments: [${runArgs}]"
+
+  properties=$(cat "${config}" | jq -j '.properties // empty | join(" ")')
+  properties="${properties//\$\{PWD\}/${PORTABLE_PWD}}"
+  properties="${properties//\$\{SCENARIO\}/${scenarioName}}"
+  properties="${properties//\%\{random\}/${random}}"
+  if [[ "$DEBUG" == "debug" ]]; then
+    printyellow "Enabling debugger on port $DEBUG_PORT"
+    runArgs="${runArgs} -Xrunjdwp:transport=dt_socket,address=$DEBUG_PORT,server=y,suspend=$DEBUG_SUSPEND"
+  fi
+  echo -e "\nLaunching CAS with properties [${properties}], run arguments [${runArgs}] and dependencies [${dependencies}]"
+  java ${runArgs} -jar "$PWD"/cas.${projectType} ${properties} \
+    -Dcom.sun.net.ssl.checkRevocation=false \
+    --spring.profiles.active=none --server.ssl.key-store="$keystore" &
+  pid=$!
+  printgreen "\nWaiting for CAS under process id ${pid}"
+  until curl -k -L --output /dev/null --silent --fail https://localhost:8443/cas/login; do
+      echo -n '.'
+      sleep 1
+  done
+  printgreen "\n\nReady!"
 fi
-echo -e "\nLaunching CAS with properties [${properties}], run arguments [${runArgs}] and dependencies [${dependencies}]"
-java ${runArgs} -jar "$PWD"/cas.${projectType} ${properties} \
-  -Dcom.sun.net.ssl.checkRevocation=false \
-  --spring.profiles.active=none --server.ssl.key-store="$keystore" &
-pid=$!
-printgreen "\nWaiting for CAS under process id ${pid}"
-until curl -k -L --output /dev/null --silent --fail https://localhost:8443/cas/login; do
-    echo -n '.'
-    sleep 1
-done
-printgreen "\n\nReady!"
-  
+
+
 clear
 scriptPath="${scenario}/script.js"
 echo -e "*************************************"
@@ -180,21 +235,22 @@ exitScript="${exitScript//\$\{SCENARIO\}/${scenarioName}}"
 
 printgreen "Done!\n"
 
-if [[ "${CI}" != "true" ]]; then
-  printgreen "Hit enter to cleanup scenario ${scenario} that ended with exit code $RC \n"
-  read -r
+if [[ "${RERUN}" != "true" ]]; then
+  if [[ "${CI}" != "true" ]]; then
+    printgreen "Hit enter to cleanup scenario ${scenario} that ended with exit code $RC \n"
+    read -r
+  fi
+
+  printyellow "\nKilling CAS process ${pid} ..."
+  kill -9 $pid
+  printyellow "\nRemoving previous build artifacts ..."
+  rm "$PWD"/cas.${projectType}
+  rm "$PWD"/ci/tests/puppeteer/overlay/thekeystore
+  rm -Rf "$PWD"/ci/tests/puppeteer/overlay
+
+  if [[ "${CI}" == "true" ]]; then
+    docker stop $(docker container ls -aq) >/dev/null 2>&1 || true
+    docker rm $(docker container ls -aq) >/dev/null 2>&1 || true
+  fi
 fi
-
-printyellow "\nKilling CAS process ${pid} ..."
-kill -9 $pid
-printyellow "\nRemoving previous build artifacts ..."
-rm "$PWD"/cas.${projectType}
-rm "$PWD"/ci/tests/puppeteer/overlay/thekeystore
-rm -Rf "$PWD"/ci/tests/puppeteer/overlay
-
-if [[ "${CI}" == "true" ]]; then
-  docker stop $(docker container ls -aq) >/dev/null 2>&1 || true
-  docker rm $(docker container ls -aq) >/dev/null 2>&1 || true
-fi
-
 exit $RC
